Add a sysctl to query s0ix preference
S0ix is the name of Intel's implementation of Low Power Idle states as
described in ACPI spec via LPIT, and SPMC. It can be thought of as a
hardware managed suspend state whereby all software needs to do is
sufficiently idle the system, and hardware will go into the lower power
state until woken by a specified set of events.

Newer laptops with Intel CPUs prefer a sleep state known as s0ix (where
x can be 1, 2, or 3). s0ix entry and benefits will be explained in
future patches. The most basic criteria to determine if a platform
supports s0ix is bit 21 of the FADT (documented here:
http://www.uefi.org/sites/default/files/resources/Intel_ACPI_Low_Power_S0_Idle.pdf).

Unfortunately BIOS are buggy, and the next couple of patches will add to
this query.