acpi/lpit: Add basic support for the LPIT
The Low Power Idle table is an ACPI table which describes the various
low power states of the platform. It's primary use is to be able to read
out residencies of the various states. With this, an interface can be
provided back to userspace and allow unprivileged userspace to ascertain
a variety of power information.

A subsequent patch to turbostat
(https://svnweb.freebsd.org/ports/head/sysutils/turbostat/) will
extrapolate the information.