acpi/spmc: Add support for the SPMC
The System Power Management Controller is a system device defined in
ACPI spec whose purpose is to provide hooks for Operating
System-directed configuration and Power Management (OSPM) to help the
platform achieve higher efficiency. The corollary to this is that the
presence of an SPMC and certain functions makes it a good candidate for
belong to a platform which supports s0ix.

This is one of the only bits that need frobbing in addition to suspend
to idle in order to make s0ix work well.

While here, add a sysctl to dump state:
sysctl hw.acpi.SPMC.min_dstate