acpi/sleep: remove unnecessary EARLY_AP sleep code
Whether or not EARLY_AP_STARTUP is used, by the time the system is
actually trying to go to sleep, smp must have been started, or there
must be only one CPU.

This also seems to fix a minor problem - an unnecessary bind/unbind when
EARLY_AP_STARTUP is defined and mp_ncpus == 1 (I assume that's weird but
possible).

The purpose of this is just code cleanup.