Turn on emulated S3 with s0ix
This feature is equivalent to using s2idle for mem_sleep in Linux, and
can be though of as functionally the same thing.

s0ix is Intel's name for a hardware feature that implements Low Power
Idle as described in ACPI spec collateral documents. It is automatically
achieved when all devices and all cores are in the requisite low power
states. Hardware will transition independently between from s0, to s0i1
and finally to s0i3. S0i3 should save roughly the same amount of power
as S3 but doesn't require calling into ACPI to transition states.

This is useful for systems which support s0ix but:
- don't have S3 at all
- have buggy S3
- have slow S3 transitions