Add a suspend to idle option (S0IDLE)
Suspend to idle is the ability to transition a system to being
completely idle with all devices powered down, but not actually in S3.
I'm dubbing this an S0 idle state. As far as the system is concerned, it
is in S0, but the OS will explicitly bring down everything else as
forcefully as necessary to resemble an S3 state.

This may be useful for systems where:
- BIOS lacks S3 support, or has flaky S3 support.
- BIOS transitions in and out of S3 very slowly.
- Full disabling of ACPI is required due to other bugs.

Similar functionality can be achieved from userspace by running devctl
suspend on all devices and having the system being completely idle. The
advantages over userspace methods are:
- Forced idleness (no callouts or the like can wake things up)
- All interrupts are disabled
- Power button and lid events cause wakeup

In addition to these benefits, it also provides a way of testing the
upcoming s0ix support in a more broad manner. While not every machine
supports s0ix, they all support suspend to idle, and the code paths are
almost identical.

In order to use this, either set the variable at runtime, or put it in
your /etc/sysctl.conf

sysctl hw.acpi.suspend_state=S0IDLE
hw.acpi.suspend_state: NONE -> S0IDLE

sysctl hw.acpi.supported_sleep_state
hw.acpi.supported_sleep_state: S5 S0IDLE

As of this patch it is only supported currently for amd64 and i386, but
that is an artificial restriction. It is also a global flag for
replacing specifically S3. That too is an artificial restriction.

I've decided to keep all of the existing suspend state infrastructure in
place and use a global override knob for S0IDLE. For the sake of
limiting damage, the next patches will attempt to allow overriding all
states to S0IDLE instead of this one.

For "compatibility" with Linux, S2IDLE is also a supported name for
this.