acpi: Combine error handling for enter/req sleep
The only real difference between these functions is that the Request
returns an internal errno while Enter is expected to return an ACPI
errno. The more minor difference is one prints the other doesn't. By
consolidating these, not only do we get equal behavior when Enter might
be called without Req, but, it will help reduce churn when this logic
becomes a little more convoluted for both cases.