acpi/wake: Move ACPI button status to wake event
With upcoming patches, we're going to want a way to suspend the system
and use the power button potentially to resume. That path will not
follow the S3 sleep path. As a result, it's important to clear the
status outside of the s3 path.

This seems more correct to me as it's possible to go through resume
without the power button being involved (like a lid event).