cpufreq: Rename hwpstate to hwpstate_amd.
That's what it is, and that's what it does. This technology is similar
to earlier Intel technologies which weren't named hwpstate. AMD allows
transitioning between pstates to be done by hardware after a request is
made by the operating system.

What Intel calls hwpstates are more autonomous and require little
intervention from the operating system.

v2: Add this to UPDATING (cem)