cpufreq: Create an uncached flag
This flag will tell the cpufreq core to get the actual frequency value
from the driver instead of trying to do tricks. In cases where a driver
cannot obtain it, it falls back to using est to estimate it.

This is useful for frequency controllers that are managed by hardware.
In such a case, software shouldn't set frequency (although in Intel's
case it's possible by setting desired = min = max), and obtaining
frequency cannot reasonably be cached.

What one could do instead of this is register a handler for frequency
change events and cache the values then, so reads won't hurt so bad.

v2: Add assertion for lock when doing estimation (cem)