cpufreq: Allow cpufreq to act over individual cpus
Prior to this patch, the cpufreq interface allows setting frequency for
all cpus, but obtaining frequency information assumes all cpus are the
same and therefore will only find the first one and return that. This
patch will register the actual cpufreq device driver registering itself,
and then prioritize that when trying to obtain frequency and levels.
With this patch the old behavior will still be fallen back to upon
errors.

NB: I don't fully understand the motivation for the original behavior