cpufreq: Add sysctl to show which freq driver
Currently, outside of dmesg, there is no way to determine which
cpu frequency driver is being used. This patch provides a new sysctl to
display the value

All cpus will use the same driver currently, and there is a 1:1
relationship between a cpu and a cpufreq driver. It isn't strictly
necessary, and so this patch provides the flexibility to do per cpufreq
displays.

Example: $ dev.cpufreq.0.freq_driver
dev.cpufreq.0.freq_driver: est